+++ lib/dealog_backoffice_web/live/approvals_live.ex
@@ -1,9 +1,0 @@
-defmodule DealogBackofficeWeb.ApprovalsLive do
-  use DealogBackofficeWeb, :live_view
-
-  @impl true
-  def mount(_params, _session, socket) do
-    socket = assign(socket, page_title: gettext("Approvals"), active_page: :approvals)
-    {:ok, socket}
-  end
-end